# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión — Proyecto Integrado",
...
        "contact": {
            "name": "Equipo de desarrollo",
            "email": "david.marquez.cordoba@gmail.com, barbaracolomer@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar todos los productos",
                "description": "Devuelve un array con todos los productos y sus categorías asociadas. Endpoint público.",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Lista de productos recuperada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "producto_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Fujifilm X-T5"
                                            },
                                            "precio": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1899.99
                                            },
                                            "cantidad": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "foto": {
                                                "type": "string",
                                                "example": "http://.../foto.jpg"
                                            },
                                            "categoria_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "categoria": {
                                                "description": "Datos de la categoría asociada",
                                                "properties": {
                                                    "categoria_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Cámaras"
                                                    }
                                                },
                                                "type": "object",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "description": "Crea un nuevo producto en la base de datos. **Solo accesible para roles de Almacén o Administrador.**",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio",
                                    "cantidad",
                                    "codigo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Tarjeta Sd"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": 89.9
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "codigo": {
                                        "type": "string",
                                        "example": "LPM-002"
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "foto": {
                                        "type": "string",
                                        "example": "https://example.com/tarjeta.jpg"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "tarjeta sd 4tb"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{producto_id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Mostrar un producto específico",
                "description": "Devuelve la información detallada de un producto y su categoría asociada.",
                "operationId": "8aa1ea210d4fa1e20a5872734e4fdc8c",
                "parameters": [
                    {
                        "name": "producto_id",
                        "in": "path",
                        "description": "ID del producto a consultar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "producto_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Fujifilm X-T5"
                                        },
                                        "precio": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1899.99
                                        },
                                        "foto": {
                                            "type": "string",
                                            "example": "http://.../foto.jpg"
                                        },
                                        "categoria_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "categoria": {
                                            "description": "Objeto con la información de la categoría",
                                            "properties": {
                                                "categoria_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Cámaras"
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto",
                "description": "Permite modificar los datos de un producto existente. **Solo accesible para roles de Almacén o Administrador.**",
                "operationId": "464eb73e3b797bf82e77553a3a897a75",
                "parameters": [
                    {
                        "name": "producto_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Lámpara LED actualizada"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": 59.99
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "foto": {
                                        "type": "string",
                                        "example": "https://example.com/lampara-nueva.jpg"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Lámpara con nuevo diseño y más brillo."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado correctamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "description": "Elimina un producto del catálogo. **Solo accesible para roles de Almacén o Administrador.**",
                "operationId": "1b3d844112ed95332546dc1198504660",
                "parameters": [
                    {
                        "name": "producto_id",
                        "in": "path",
                        "description": "ID del producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación por token Bearer (Laravel Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}